+++ CHANGELOG.md
@@ -79,6 +79,7 @@
 - **Breaking Change**: rename builtin ``eval`` command with ``call``
     - ``eval`` is still builtin command for future usage
 - **Breaking Change**: invalid utf8 bytes are always grapheme/word boundary
+- **Breaking Change**: in finally/defer block, does not ignore exceptions that can be caught within finally/defer block
 - show stack trace of ignored exceptions within finally/defer block
 - improve some semantic error messages
 - allow ``Nothing?`` type